// apps/pwabuilder/Frontend/src/script/utils/environment.d.ts
export declare const env: {
    isProduction: boolean;
    manifestCreatorUrl: string;
    imageProxyUrl: string;
    api: string;
    windowsPackageGeneratorUrl: string;
    androidPackageGeneratorUrl: string;
    iosPackageGeneratorUrl: string;
    safeUrlFetcher: string;
    validateGiveawayUrl: string;
    tokensCampaignRunning: boolean;
};